Optimize how the program counter works, shaves off 0.2s for arith.lua

+++ src/opcode.h
@@ -56,7 +56,7 @@
 #define B(opc)        (((opc) >> 23) & 0x1ff)
 #define C(opc)        (((opc) >> 14) & 0x1ff)
 #define PAYLOAD(opc)  (((opc) >> 14) & 0x3ffff)
-#define UNBIAS(v)     ((v) - 131071)
+#define UNBIAS(v)     ((i32) ((v) - 131071))
 
 void opcode_dump(FILE *out, uint32_t code);
 void opcode_dump_idx(FILE *out, lfunc_t *func, size_t idx);
